# Conflicts:
#	src/main/java/com/sparta/hotbody/user/service/KakaoService.java
#	src/main/resources/static/login.html

+++ src/main/java/com/sparta/hotbody/post/repository/PostLikeRepository.java
@@ -9,5 +9,5 @@ public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
 
   void deleteByPostIdAndUserId(Long postId, Long userId);
 
-  Long countAllByPostId(Long postId);
+  int countAllByPostId(Long postId);
 }
